add file upload for avatar, video, design webpack and scss

+++ src/controllers/videoController.js
@@ -1,8 +1,8 @@
-import Video from '../models/Video';
-
+import Video from "../models/Video";
+import User from "../models/User";
 const handleSearch = (error, videos) => {
-  console.log('error', error);
-  console.log('video', videos);
+  console.log("error", error);
+  console.log("video", videos);
 };
 /*
 console.log("start")
@@ -13,63 +13,83 @@ console.log("finished")
 */
 
 export const home = async (req, res) => {
-  const videos = await Video.find({}).sort({ createdAt: 'desc' });
+  const videos = await Video.find({}).sort({ createdAt: "desc" }).populate("owner");
   // console.log(videos);
 
-  return res.render('home', { pageTitle: 'Home', videos });
+  return res.render("home", { pageTitle: "Home", videos });
 };
 export const watch = async (req, res) => {
   const { id } = req.params;
   //const id =req.params.id;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
 
   if (!video) {
-    return res.status(404).render('404', { pageTitle: 'Video is not found' });
+    return res.status(404).render("404", { pageTitle: "Video is not found" });
   }
-  return res.render('watch', { pageTitle: video.title, video });
+  return res.render("watch", { pageTitle: video.title, video });
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
-    return res.status(404).render('404', { pageTitle: 'Video is not found' });
+    return res.status(404).render("404", { pageTitle: "Video is not found" });
   }
-  return res.render('edit', { pageTitle: `Edit!!! ${video.title}`, video });
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("edit", { pageTitle: `Edit!!! ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
 
   const video = await Video.exists(id);
   if (!video) {
-    return res.status(404).render('404', { pageTitle: 'Video is not found' });
+    return res.status(404).render("404", { pageTitle: "Video is not found" });
   }
-  /*video.title = title;video.description = description; video.hashtags = hashtags.split(',').map((word) => (word.startsWith('#') ? word : `#${word}`));
-  await video.save(); */
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndUpdate(id, { title, description, hashtags: Video.formatHashtags(hashtags) });
   return res.redirect(`/videos/${id}`);
 };
 
 export const getUpload = (req, res) => {
-  return res.render('upload', { pageTitle: 'Upload Video' });
+  return res.render("upload", { pageTitle: "Upload Video" });
 };
 export const postUpload = async (req, res) => {
   // here we will add a video to the videos array
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
 
   /* const video = new Video;
      await video.save();  ->> await Video.create  makes easy and once. */
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
-    return res.redirect('/');
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
+    return res.redirect("/");
   } catch (error) {
-    return res.status(400).render('upload', {
-      pageTitle: 'Upload Video',
+    return res.status(400).render("upload", {
+      pageTitle: "Upload Video",
       errorMessage: error._message,
     });
   }
@@ -77,10 +97,19 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
-  console.log(id);
-  //delete video
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video is not found" });
+  }
+
   await Video.findByIdAndDelete(id);
-  return res.redirect('/');
+  return res.redirect("/");
 };
 
 export const search = async (req, res) => {
@@ -90,9 +119,9 @@ export const search = async (req, res) => {
     //search
     videos = await Video.find({
       title: {
-        $regex: new RegExp(`${keyword}*`, 'i'),
+        $regex: new RegExp(`${keyword}*`, "i"),
       },
-    });
+    }).populate("owner");
   }
-  return res.render('search', { pageTitle: 'Search', videos });
+  return res.render("search", { pageTitle: "Search", videos });
 };
